move fakeNodeCreator under k8s mocks
